Add files via upload
This is my final project for General Assembly. 

The task was to deliver a panda project written in a Jupiter Notebook + a 5 minutes presentation in class. We could get the data set of our likening and then clean it, check for errors/empty/invalid values, make hypothesis and answer then using everything we have learn in class.